fix only space in input bug

+++ lexer/lexer.c
@@ -6,7 +6,7 @@
 /*   By: lray <lray@student.42lausanne.ch >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/09 23:26:56 by lray              #+#    #+#             */
-/*   Updated: 2023/08/11 23:27:07 by lray             ###   ########.fr       */
+/*   Updated: 2023/08/12 00:03:17 by lray             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,8 +62,8 @@ static t_dyntklist	*lexer_run(char *input)
 
 static int	is_only_space(char *input)
 {
-	size_t	nbr_space;
-	size_t	i;
+	int	nbr_space;
+	int	i;
 
 	i = 0;
 	nbr_space = 0;
